+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 1.16 (pre-release):
+* [Fix #27] Allow light color and status be controllable in EVA and in flight.
 * [Change] Stop complaining about KSP minor version change.
 
 # 1.15 (December 1st, 2019):